[Security Solution] Fix alerts table potentially not applying alert assignees (#219460)
## Summary

This PR implements a similar change that was just merged a few hours
ago. While [that change](https://github.com/elastic/kibana/pull/219410)
was made to the alert tags not always working on the alerts table, this
current change is applied to the alert assignees that faced a potential
similar issue. The alert assignee code was introduced in [this
PR](https://github.com/elastic/kibana/pull/170579), and I believe the
code was using the similar logic of [the alert tag
PR](https://github.com/elastic/kibana/pull/157786).

The issue is related to the fact that we have a `useRef` for a function
that is returned before the `useEffect` in the same hook runs, and
setting the value of the function returned is happening within that
`useEffect`. This has not caused any issues because the few places where
this code is being used (the alerts page alerts table) is extremely not
efficient and renders multiple times. This gives enough tries to the
hook to actually get a value and return the correct function.

This PR fixes that by returning the function directly.

Here's a video showing that the functionality still works correctly for
bulk actions:


https://github.com/user-attachments/assets/b3394ffe-8333-4e0a-9bf7-831ef8ea8aea

And also for normal row actions:


https://github.com/user-attachments/assets/5f8c9d23-f0ef-4c65-b7de-4dc34478a8e7

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios