Fix ignored dynamic templates (#219875)
This PR fixes the bug introduced with:
https://github.com/elastic/kibana/pull/216719

We didn't pass the dynamic_templates param while creating the component
template.
The fields were still being added because the fields were marked as
`dynamic: true`

As the dynamic_template was ignored, the fields were added with an extra
mapping like `filedname.keyword`
This PR fixes that too.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>