[Lens] [ES|QL] Tagcloud missing from types (#219916)
## Summary

Closes https://github.com/elastic/kibana/issues/219905

We are using the `mapVisToChartType` to return the lens visualization
type. If it is not in the list then it will return undefined

```
const mapVisToChartType = (visualizationType: string) => {
  if (isLensVisualizationType(visualizationType)) {
    return lensTypesToChartTypes[visualizationType];
  }
```

This was a suggestion from the viz team but because it is not used
across the Lens code we missed the tagcloud. Using it across Lens will
ensure that this won't happen again
https://github.com/elastic/kibana/issues/219917.

This PR just adds the tagclouds in the list to ensure that the bug is
fixed.



![meow](https://github.com/user-attachments/assets/1aeb96ac-c41b-4d14-9654-30573b548878)