Fixing layout of the embeddablePanelHeading containing description an… (#219428)
It has been assumed that everything of the embeddable panel heading will
stay on one line, and this is a proposal based on that assumption.

This video shows how the layout responds:


https://github.com/user-attachments/assets/db7f3e38-1fbd-4d21-a0a2-96ed1768b744



Introduced ellipsis for panel titles and custom time range badges when
the text overflows.

Fixed the issue with the description icon, which was wrapping to the
next line.

Fixed the layout of the embeddable panel heading containing three
elements: title, icon, and badge.

Closes: #211252 and #219868


Improved layout of children within the parent container, as the previous
styles didn’t prevent significant disproportion caused by varying title
and time-range-badge text lengths (disproportion is shown in the first
screenshot below).
![Screenshot 2025-05-02 at 19 39
01](https://github.com/user-attachments/assets/23c4a4ae-2dc0-4c04-9ecf-ed4abfba8c76)
Now, each child (which is title and time-range-badge) will grow/shrink
proportionally, ensuring consistent layout regardless of their content
length. (as shown in the screenshots below)
![Screenshot 2025-05-02 at 19 39
30](https://github.com/user-attachments/assets/a48d1c9e-2bb0-4c55-96be-3b379091e904)
![Screenshot 2025-05-02 at 19 40
06](https://github.com/user-attachments/assets/72171174-6565-412a-abb9-1fc53b7b094c)

---------

Co-authored-by: Hannah Mudge <Heenawter@users.noreply.github.com>
Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>