🌊 Streams: Fix responsiveness bugs in routing UI (#220742)
Noticed a couple of small problems while coding on a 13 inch screen and
decided to fix this. I don't think the smaller viewports are super
important for us and I don't think we need to shoot for a super polished
mobile experience, but these were easy to do and make using the UI on
smaller screens a much better experience.

Didn't go all the way to make it nice, if we think it's important we can
properly prioritize it, this is an 80/20 thing.

* Add `wrap={true}` to horizontal tool bars that can get pretty small on
mobile screens
* Disable automatic viewport-based wrapping in cases where this doesn't
help
* Switch to a vertical switch on smaller screens
* Reduce the min size of the resizable panel, because sometimes it goes
under the min size anyway on small screens which causes it to wiggle
when hovering over the entries