[Oblt Onboarding][Auto Detect] Handle Agent's DEGRADED state (#220527)
Closes https://github.com/elastic/observability-dev/issues/4505

This change updates the logic for Auto Detect flow to handle `DEGRADED`
state reported by Elastic Agent.

## Context
Previously, if any of the the configured data streams were in a degraded
state, we would report to the user that the flow has failed even though
some data streams might be healthy and even send the data, and Kibana in
turn would never check for incoming data and wouldn't report any
additional telemetry events.

Additionally, if user installs any of the integrations that require
further configuration (e.g. MySQL or Postgres) they would be guarantied
to be in a degraded state because Fleet API provides configuration for
metrics data streams and they require login and password.

This change makes handling degraded state more optimistic. If there is
at least one data stream in a `HEALTHY` state, we're now reporting a
successful status of the whole flow to the user. Additional, we're
showing status of the unhealthy data stream as a separate message if
there are any (see screenshots). We report the whole flow as failed only
when none of the data streams are healthy.

| All data streams healthy | Some data streams degraded |
| --- | --- |
| ![CleanShot 2025-05-08 at 16 35
55@2x](https://github.com/user-attachments/assets/8ee8a694-46a7-4e8f-9694-a2d0dc61b72f)
| ![CleanShot 2025-05-08 at 16 38
10@2x](https://github.com/user-attachments/assets/58c493ab-f5ac-4e5f-80b6-5c135711e251)
|


## How to test

Go through the Auto Detect flow once with integrations that don't
require additional configuration (e.g. System, Nginx) and another time
with something like MySQL or Postgres. Make sure that everything works
as expected and messages from the script make sense.

Co-authored-by: Joe Reuter <johannes.reuter@elastic.co>